--- Assets/Scripts/ChooseStarButton.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6dba1027863946dbb4e3001fd307f5f5
-timeCreated: 1584572501
