+++ Assets/Scripts/StarIconManager.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 8d35f40dd23c4d1c90f9804f16142293
-timeCreated: 1584588617
+guid: 6dba1027863946dbb4e3001fd307f5f5
+timeCreated: 1584572501
